requirements.txt
typing-extensions==3.10.0
python-dateutil==2.8.2
packaging==21.2
flatbuffers==23.1.21
requests==2.31.0
chardet==4.0.0
elastic-transport==8.0.0
google-auth==2.15.0
protobuf==3.19.6
urllib3==1.26.2
grpcio==1.48.2
testresources
numpy==1.23.5
setuptools
scipy
scikit-learn==0.20.3
opencv-python==4.2.0.32
opencv-contrib-python==4.2.0.32
tensorflow-gpu==2.11.0
keras_applications
tensorflow-model-optimization==0.5.0
tensorflow-addons
matplotlib
tqdm
pillow
mnn
Cython
pycocotools
keras2onnx
tf2onnx==0.4.2
onnx
onnxruntime
tfcoreml==1.1
sympy
imgaug
imagecorruptions
bokeh==2.4.0
tidecv
pyproject-toml
tensorrt
